fixed a silly mistake on enumerating the vertexes

+++ filtration_v02/enumerate_nodes_edges_by_integers.c
@@ -19,7 +19,7 @@ typedef struct {
 typedef struct {
     node         from;
     node         to;
-    unsigned int weight;
+    unsigned long int weight;
 
     char         flag;
